requirements.txt
# AI Providers
groq
openai
anthropic
google-generativeai
cohere

# Flask
flask
flask-cors

# Yardımcı Kütüphaneler
python-dotenv
colorama
markdown
pygments
gtts


# Flask Web Server
flask==3.0.0
flask-cors==4.0.0

# AI Providers
groq==0.4.2
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2
cohere==4.47

# Web Search
tavily-python==0.3.3

# Text to Speech
gTTS==2.5.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Database
sqlite3  # Built-in Python